Seperate out map and stats to seperate pages
