+++ src/app/app-routing.module.ts
@@ -6,6 +6,7 @@ import { HomeComponent } from './home/home.component';
 import { CheckinsComponent } from './checkins/checkins.component';
 import { TopbeersComponent } from './topbeers/topbeers.component';
 import { BadgesComponent } from './badges/badges.component';
+import { MapComponent } from './map/map.component';
 import { StatsComponent } from './stats/stats.component';
 import { WishlistComponent } from './wishlist/wishlist.component';
 import { AboutComponent } from './about/about.component';
@@ -16,6 +17,7 @@ export const routes: Routes = [
   { path: 'checkins', component: CheckinsComponent },
   { path: 'topbeers', component: TopbeersComponent },
   { path: 'badges', component: BadgesComponent },
+  { path: 'map', component: MapComponent },
   { path: 'stats', component: StatsComponent },
   { path: 'wishlist', component: WishlistComponent },
   { path: 'about', component: AboutComponent }
