+++ src/app/app.component.html
@@ -39,6 +39,13 @@
       </button>
       <button
         mat-button
+        routerLink="/map"
+        routerLinkActive="router-link-active no-border"
+      >
+        Breweries Map
+      </button>
+      <button
+        mat-button
         routerLink="/stats"
         routerLinkActive="router-link-active no-border"
       >
@@ -82,6 +89,15 @@
           fxLayoutAlign="start"
         >
           <mat-icon>badge</mat-icon>Badges
+        </button>
+        <mat-divider></mat-divider>
+        <button
+          mat-button
+          routerLink="/map"
+          routerLinkActive="selected no-border txt-bold"
+          fxLayoutAlign="start"
+        >
+          <mat-icon>map</mat-icon>Map
         </button>
         <mat-divider></mat-divider>
         <button
