+++ src/app/map/map.component.html
@@ -1,0 +1,5 @@
+<div class="map-container">
+  <div class="map-frame">
+    <div id="map"></div>
+  </div>
+</div>
