+++ src/app/stats/stats.component.html
@@ -1,5 +1,5 @@
-<div class="map-container">
-  <div class="map-frame">
-    <div id="map"></div>
-  </div>
+<h1 mat-dialog-title>Stats</h1>
+<div class="wrapper">
+  <h2>Coming Soon<span class="dot">.</span></h2>
+    <p>Pardon our dust.</p>
 </div>
