+++ src/app/stats/stats.component.ts
@@ -1,50 +1,15 @@
-import { AfterViewInit, Component } from '@angular/core';
-import { MarkerService } from '../_services/marker.service';
-import * as L from 'leaflet';
-
-const iconRetinaUrl = '/assets/images/marker-icon-2x.png';
-const iconUrl = 'assets/images/marker-icon.png';
-const shadowUrl = 'assets/images/marker-shadow.png'
+import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-stats',
   templateUrl: './stats.component.html',
   styleUrls: ['./stats.component.css']
 })
-export class StatsComponent implements AfterViewInit {
+export class StatsComponent implements OnInit {
 
-  private map: any;
-  private states: any;
+  constructor() { }
 
-  private markerIcon = L.icon({
-    iconRetinaUrl,
-    iconUrl,
-    shadowUrl,
-    iconSize: [25, 41],
-    iconAnchor: [12, 41],
-    popupAnchor: [1, -34],
-    tooltipAnchor: [16, -28],
-    shadowSize: [41, 41]
-  });
-
-  constructor(private markerService: MarkerService,) {}
-
-  ngAfterViewInit(): void {
-    this.initMap();
-    this.markerService.makeBreweryMarkers(this.map, this.markerIcon);
+  ngOnInit(): void {
   }
 
-  private initMap(): void {
-    this.map = L.map('map', {
-      center: [39.8282, -98.5795],
-      zoom: 3
-    });
-
-    const tiles = L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {
-      maxZoom: 19,
-      attribution: '&copy; <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a>'
-    });
-
-    tiles.addTo(this.map);
-  }
 }
