+++ src/styles.css
@@ -488,3 +488,5 @@ body {
     width: 50%;
   }
 }
+html, body { height: 100%; }
+body { margin: 0; font-family: Roboto, "Helvetica Neue", sans-serif; }
